# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API к ИС Не-Ной",
        "title": "Ne-Noy-Api",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://simply-funny-spearfish.cloudpub.ru",
    "basePath": "/api",
    "paths": {
...
        "/v1/events/test": {
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Создать тест",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для создания теста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_dto.CreateTestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/test_dto.TestDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/events/test/{id}": {
            "get": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Получить тест по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/test_dto.TestDto"
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Тест не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Удалить тест",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Тест не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Обновить тест",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления теста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_dto.UpdateTestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/test_dto.TestDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Тест не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/events/test/{id}/q": {
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Добавить вопрос в тест",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные вопроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_dto.AddQuestionDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/test_dto.QuestionDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Тест не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/events/test/{id}/q/{qId}": {
            "get": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Получить вопрос теста",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID вопроса",
                        "name": "qId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/test_dto.QuestionDto"
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "VkAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Сохранить ответ пользователя на вопрос",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор запроса",
                        "name": "X-Request-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID вопроса",
                        "name": "qId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ответ пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_dto.SetAnswerDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/test_dto.UserAnswerDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Вопрос или вариант ответа не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "test_dto.AddQuestionDto": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "test_dto.CreateTestDto": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "ext_link_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "vk_post_id": {
                    "type": "integer"
                }
            }
        },
...
        "test_dto.SetAnswerDto": {
            "type": "object",
            "properties": {
                "answer_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "test_dto.TestDto": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "ext_link_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/test_dto.QuestionDto"
                    }
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "vk_post_id": {
                    "type": "integer"
                }
            }
        },
...
        "test_dto.UpdateTestDto": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "cover": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "ext_link_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "vk_post_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "VkAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}